BUG/MEDIUM: listener/proxy: fix listeners notify for proxy resume (2nd try)
In 7741928c ("BUG/MEDIUM: listener/proxy: fix listeners notify for proxy
resume"), we changed the way we detect that the proxy is resumed by
checking the suspend counter instead of the paused counter to take into
account listeners that don't support being paused such as abns socket.

However, the backport note for the 2.4 version was wrong. Since we don't
have the PR_FL_PAUSED flag, we must still ensure that we report the
change only if the proxy was previously suspended. It can be done by
comparing the new state with the previous state like it was done prior to
7741928c.

This bug was raised by Sebastien: it can be easily reproduced when proxy
rate-limiting directives such as 'maxconnrate' are used and the proxy
becomes limited and is then relaxed when the rate becomes acceptable
again. Due to this, a lot of polluting logs could be generated:

   [WARNING] (2661577) : Resumed frontend GLOBAL.
   [WARNING] (2661577) : Resumed frontend GLOBAL.
   [WARNING] (2661577) : Resumed frontend GLOBAL.

This is a 2.4 only fix, no backport needed.