MINOR: ext-check: add an option to preserve environment variables
In Github issue #2128, @jvincze84 explained the complexity of using
external checks in some advanced setups due to the systematic purge of
environment variables, and expressed the desire to preserve the
existing environment. During the discussion an agreement was found
around having an option to "external-check" to do that and that
solution was tested and confirmed to work by user @nyxi.

This patch just cleans this up, implements the option as
"preserve-env" and documents it. The default behavior does not change,
the environment is still purged, unless "preserve-env" is passed. The
choice of not using "import-env" instead was made so that we could
later use it to name specific variables that have to be imported
instead of keeping the whole environment.

The patch is simple enough that it could be backported if needed (and
was in fact tested on 2.6 first).

(cherry picked from commit 1de44daf7d61828be1effd970817ee59ed8750d7)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>
(cherry picked from commit 39dcd1fffa5dd2a11e1247c6154646d0ee440c33)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>
(cherry picked from commit ae9418aca79cd358b8ac5681453024ff58e96c37)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>