BUG/MINOR: quic: Packet number spaces too lately initialized
This bug could be reproduced with -dMfail option. When a new QUIC connection
could not be allocated because of a failed <buf_area> allocation, this could
lead quic_conn_release() to be called without any packet number space initialized.
Then qc_release_pktns_frms() is called and derefences non initialized pointers.

To fix this, initialize the packet number spaces before allocation <buf_area>.

This bug is specific to 2.8.