BUG/MEDIUM: stconn: Forward shutdown on write timeout only if it is forwardable
The commit b9c87f8082 ("BUG/MEDIUM: stconn/stream: Forward shutdown on write
timeout") introduced a regression. In sc_cond_forward_shut(), the write
timeout is considered too early to forward the shutdown. In fact, it is
always considered, even if the shutdown is not forwardable yet. It is of
course unexpected. It is especially an issue when a write timeout is
encountered on server side during the connection establishment. In this
case, if shutdown is forwarded too early on the client side, the connection
is closed before the 503 error sending.

So the write timeout must indeed be considered to forward the shutdown to
the underlying layer, but only if the shutdown is forwardable. Otherwise, we
should do nothing.

This patch should fix the issue #2404. It must be backported as far as 2.2.

(cherry picked from commit 7eb7ae283557d8467403b9db35605861a0870add)
Signed-off-by: Willy Tarreau <w@1wt.eu>
(cherry picked from commit 2b52339ac0fefd9561372e7a777f098dfa00ef1d)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>
(cherry picked from commit c97832701ec5c606a2c3c2b383617bbbff1efaa5)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>